auth-service/requirements.txt
fastapi
uvicorn
pydantic
passlib
bcrypt
python-jose[cryptography]
python-multipart

